requirements.txt
numpy~=1.25.2
pandas~=2.0.3
python-box~=7.0.1
pyvroom~=1.13.2
boltons~=23.0.0
loguru~=0.7.0
pydantic~=2.1.1
more-itertools~=10.0.0
wrapt~=1.15.0
pytest~=7.4.0